A new approach to render layers
Now, layer render to the buffer canvas, which size is larger than the size of visible GUI canvas. Visible canvas displays part of buffer canvas. Rendering happens only when data changes, not when a user scrolls the camera. This approach avoids render during mouse move, which is more performant, especially in the Firefox.